+++ udata_front/theme/gouvfr/datagouv-components/src/components/AsyncTextClamp/AsyncTextClamp.vue
@@ -1,0 +1,18 @@
+<template>
+  <TextClamp
+    v-if="resolvedText"
+    :text="resolvedText"
+    :max-lines
+  />
+</template>
+
+<script setup lang="ts">
+import TextClamp from "vue3-text-clamp"
+
+const props = defineProps<{
+  maxLines: number
+  text: Promise<string>
+}>();
+
+const resolvedText = await props.text;
+</script>