+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix missing closing <a> tag in contact form [#589](https://github.com/datagouv/udata-front/pull/589)
 - Fix breadcrumb link for dataservice page [#590](https://github.com/datagouv/udata-front/pull/590)
+- Remove truncate and excerpt helpers function and fix error with remove-markdown [#591](https://github.com/datagouv/udata-front/pull/591)
 
 ## 6.0.0 (2024-11-07)